feat: creator "hotkeysvs" using ScreenShot

+++ src/example/exwidget.h
@@ -46,7 +46,7 @@ protected:
     virtual void paintEvent(QPaintEvent* event) override;
 
 private:
-    int m_scal;
+    double m_scal;
     Qt::Orientations m_orien;
     QGridLayout* m_layout;
     QMap<QString, QString> m_labMap;
