+++ src/tool/base/managebar.h
@@ -33,7 +33,7 @@ private:
 signals:
 
 private:
-    int m_scal;
+    double m_scal;
     Qt::Orientations m_orien;
     QBoxLayout* m_layout;
 };
